Replace the use of the file module for permission recursion with chown
I've noticed over the course of using the role myself and testing that
the use of Ansible's file module to recursively ensure permissions for
the bundle directory is quite slow and can significantly impact the time
it takes the role to complete. Chown in this instance is much, much
faster. I'm commenting out the old functionality for posterity but will
probably remove it in the future or revert back to Ansible's way if it
improves in speed.